% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce_seabird_doc.R
\docType{data}
\name{cce_seabird}
\alias{cce_seabird}
\title{Bird Community Abundance and Richness - Seasonal}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 73 rows and 4 columns.
}
\source{
{Farallon Institute Advanced Ecosystem Research, CalCOFI - Scripps Institution of Oceanography, and B. Sydeman. 2017. Seasonal seabird density and richness of seabirds off Southern California from sampling aboard California Cooperative Oceanic Fisheries Investigations (CalCOFI) cruises, 1987 - 2006. ver 3. Environmental Data Initiative. https://doi.org/10.6073/pasta/73ce2f6bf27dbf3fe7e447333a2f7e69 (Accessed 2021-03-01).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-cce.162.3}
}
\usage{
cce_seabird
}
\description{
Seasonal seabird density and richness of seabirds off Southern California from sampling aboard California Cooperative Oceanic Fisheries Investigations (CalCOFI) cruises, 1987 - 2006.
}
\keyword{datasets}
