% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce_sealevel_doc.R
\docType{data}
\name{cce_sealevel}
\alias{cce_sealevel}
\title{Add a documentation title here}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1335 rows and 8 columns.
}
\source{
{Enter dataset citation here}
\url{Add link for data package here}
}
\usage{
cce_sealevel
}
\description{
Add a subtitle / brief summary here
}
\keyword{datasets}
