% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cce_watertemp_doc.R
\docType{data}
\name{cce_watertemp}
\alias{cce_watertemp}
\title{Shore Stations Program SIO Pier Water Temperature}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36230 rows and 6 columns.
}
\source{
{Shore Stations Program. 2019. Daily sea-surface temperature measurements, collected and provided by the Shore Stations Program, La Jolla CA, 1916 - May 2019 ver 4. Environmental Data Initiative. https://doi.org/10.6073/pasta/d8e2473a3de462bd4d75bad8f934a40a (Accessed 2021-03-01).}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-cce.15.4}
}
\usage{
cce_watertemp
}
\description{
Daily sea-surface temperature measurements, collected and provided by the Shore Stations Program, La Jolla CA, 1916 - May 2019
}
\keyword{datasets}
